Renamed gtfs resource constraints
