+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/gtfs/resources/GtfsPublicTransportAgencyResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.gtfs.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.gtfs.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPublicTransportAgencyResourceContainer;
 
 @ApplicationScoped
-public class GtfsPublicTransportAgencyCache extends AbstractPublicTransportAgencyResourceContainer {
+public class GtfsPublicTransportAgencyResourceContainer extends AbstractPublicTransportAgencyResourceContainer {
 
   private static final long serialVersionUID = 5885760951047040901L;
