+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/gtfs/resources/GtfsPublicTransportRouteResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.gtfs.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.gtfs.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPublicTransportRouteResourceContainer;
 
 @ApplicationScoped
-public class GtfsPublicTransportRouteCache extends AbstractPublicTransportRouteResourceContainer {
+public class GtfsPublicTransportRouteResourceContainer extends AbstractPublicTransportRouteResourceContainer {
 
   private static final long serialVersionUID = -1804472147095708248L;
