+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/gtfs/resources/GtfsPublicTransportScheduleResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.gtfs.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.gtfs.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPublicTransportScheduleResourceContainer;
 
 @ApplicationScoped
-public class GtfsPublicTransportScheduleCache extends AbstractPublicTransportScheduleResourceContainer {
+public class GtfsPublicTransportScheduleResourceContainer extends AbstractPublicTransportScheduleResourceContainer {
 
   private static final long serialVersionUID = -5861987244184893788L;
