+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/gtfs/resources/GtfsPublicTransportStopResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.gtfs.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.gtfs.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPublicTransportStopResourceContainer;
 
 @ApplicationScoped
-public class GtfsPublicTransportStopCache extends AbstractPublicTransportStopResourceContainer {
+public class GtfsPublicTransportStopResourceContainer extends AbstractPublicTransportStopResourceContainer {
 
   private static final long serialVersionUID = 1195621948823629420L;
