+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/gtfs/resources/GtfsPublicTransportStopTimeResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.gtfs.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.gtfs.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPublicTransportStopTimeResourceContainer;
 
 @ApplicationScoped
-public class GtfsPublicTransportStopTimeCache extends AbstractPublicTransportStopTimeResourceContainer {
+public class GtfsPublicTransportStopTimeResourceContainer extends AbstractPublicTransportStopTimeResourceContainer {
 
   private static final long serialVersionUID = -889639898690519914L;
