+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/gtfs/resources/GtfsPublicTransportTripResourceContainer.java
@@ -1,11 +1,11 @@
-package fi.otavanopisto.kuntaapi.server.integrations.gtfs.cache;
+package fi.otavanopisto.kuntaapi.server.integrations.gtfs.resources;
 
 import javax.enterprise.context.ApplicationScoped;
 
 import fi.otavanopisto.kuntaapi.server.resources.AbstractPublicTransportTripResourceContainer;
 
 @ApplicationScoped
-public class GtfsPublicTransportTripCache extends AbstractPublicTransportTripResourceContainer {
+public class GtfsPublicTransportTripResourceContainer extends AbstractPublicTransportTripResourceContainer {
 
   private static final long serialVersionUID = 8900780864580591473L;
